fix help model view (#11)
* fix help page model view

* fix alignment of service and help page
