+++ src/valuable/pages/servicebanner.tsx
@@ -11,42 +11,36 @@ const ServiceBanner: React.FC<BannerProps> = ({ path }) => {
     <>
       <div className={`h-60 ${path ==='request-advanced-analytics-enviornment' ? 'bg-mso-logo' : 'bg-banner-logo'} w-300 flex  bg-center `}></div>
       <div className="w-full flex ">
-        <div className="w-[65%] max-w-7xl  relative m-auto ">
+        <div className="w-[65%] max-w-4xl  relative m-auto pl-[14rem]">
           <div className="max-w-7xl items-center justify-center ">
             {path === 'request-advanced-analytics-enviornment' ? (
               <>
                 <div className="flex items-center  justify-center ">
-                  <div className="absolute  justify-center w-full bg-white    p-10 mb-10 text-center">
+                  <div className="absolute  justify-center w-full bg-white    p-10 mb-20 text-center">
                     <h1 className="text-2xl font-bold  ">Advanced Analytics Environment(AAE)</h1>
                   </div>
-                  <img className="absolute mt-20 items-center " src={logo} alt="logo" />
+                  <img className="absolute  items-center " src={logo} alt="logo" />
                 </div>
-                <div className="px-10 py-10 ">
-                  <div className="py-3  justify ">
-                    <h2 className="font-bold py-2">
+                <div className="px-10 py-4 ">
+                  <div className="justify ">
+                    <h2 className="font-bold ">
                       Download Paragraph Completion Questions PDF for Bank Exams 2022 at Smartkeeda?
                     </h2>
                     <p>
                       Paragraph Completion is the integral part of the Verbal Reasoning Section asked in all the major
                       competitive examinations. In Paragraph Completion Questions, a paragraph is generally given by the
-                      author in which some statements will be missing which has to be answered by the students. These
-                      paragraphs are generally part of the bigger articles and revolve around a particular theme. It is
-                      important for a student to understand that theme and then answer the Questions. These Questions
-                      tests the Reading Comprehension and Vocabulary of the students.
+                      author in which some statements will be missing which has to be answered by the students.
                     </p>
 
                   </div>
-                  <div className=" py-3   justify ">
-                    <h2 className="font-bold py-2">
+                  <div className=" justify ">
+                    <h2 className="font-bold py-1">
                       Download Paragraph Completion Questions PDF for Bank Exams 2022 at Smartkeeda?
                     </h2>
                     <p>
                       Paragraph Completion is the integral part of the Verbal Reasoning Section asked in all the major
                       competitive examinations. In Paragraph Completion Questions, a paragraph is generally given by the
-                      author in which some statements will be missing which has to be answered by the students. These
-                      paragraphs are generally part of the bigger articles and revolve around a particular theme. It is
-                      important for a student to understand that theme and then answer the Questions. These Questions
-                      tests the Reading Comprehension and Vocabulary of the students.
+                      author in which some statements will be missing which has to be answered by the students.
                     </p>
 
                   </div>
@@ -59,37 +53,31 @@ const ServiceBanner: React.FC<BannerProps> = ({ path }) => {
             ): 
             path=== 'request-data-update'?(<>
             <div className="flex items-center  justify-center ">
-                  <div className="absolute  justify-center w-full bg-white    p-10 mb-10 text-center">
+                  <div className="absolute  justify-center w-full bg-white  p-10 mb-20 text-center">
                     <h1 className="text-2xl font-bold  ">Request for Data Upload</h1>
                   </div>
-                  <img className="absolute mt-20 items-center " src={logo} alt="logo" />
+                  <img className="absolute  items-center " src={logo} alt="logo" />
                 </div>
-                <div className="px-10 py-10 ">
-                  <div className="py-3  justify ">
-                    <h2 className="font-bold py-2">
+                <div className="px-10  py-4 ">
+                  <div className="justify ">
+                    <h2 className="font-bold">
                       Download Paragraph Completion Questions PDF for Bank Exams 2022 at Smartkeeda?
                     </h2>
                     <p>
                       Paragraph Completion is the integral part of the Verbal Reasoning Section asked in all the major
                       competitive examinations. In Paragraph Completion Questions, a paragraph is generally given by the
-                      author in which some statements will be missing which has to be answered by the students. These
-                      paragraphs are generally part of the bigger articles and revolve around a particular theme. It is
-                      important for a student to understand that theme and then answer the Questions. These Questions
-                      tests the Reading Comprehension and Vocabulary of the students.
+                      author in which some statements will be missing which has to be answered by the students.
                     </p>
 
                   </div>
-                  <div className=" py-3   justify ">
-                    <h2 className="font-bold py-2">
+                  <div className=" py-1 justify ">
+                    <h2 className="font-bold ">
                       Download Paragraph Completion Questions PDF for Bank Exams 2022 at Smartkeeda?
                     </h2>
                     <p>
                       Paragraph Completion is the integral part of the Verbal Reasoning Section asked in all the major
                       competitive examinations. In Paragraph Completion Questions, a paragraph is generally given by the
-                      author in which some statements will be missing which has to be answered by the students. These
-                      paragraphs are generally part of the bigger articles and revolve around a particular theme. It is
-                      important for a student to understand that theme and then answer the Questions. These Questions
-                      tests the Reading Comprehension and Vocabulary of the students.
+                      author in which some statements will be missing which has to be answered by the students.
                     </p>
 
                   </div>
